+++ compile-testing/src/main/java/com/google/testing/compile/JavaSourcesSubjectFactory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.auto.factory.gentest;
+package com.google.testing.compile;
 
 import javax.annotation.processing.Processor;
 import javax.tools.JavaFileObject;
