+++ compile-testing/src/test/resources/HelloWorld-different.java
