+++ compile-testing/src/test/resources/HelloWorld-v2.java
